Moved run actions to admin actions file
